Built-In Args
Realized that I want the arg system to work a bit differently than to how vercel/arg would support, so I'm making my own! It works similarly to FFmpeg (kind of), more that each thing you pass in is an individual input, rather than needing to specify that it's an input every time (I guess FFmpeg doesn't really do it that way, but the repetitive -i flag is kind of what I'm going for, just that each one is a different export, not all for one single export).