Overwrite Flag
Still need to set this up to work with Node's FS API calls. This is mainly to streamline usage of FFmpeg, to not silently hang the terminal window for asking whether to overwrite the file or not, because I currently don't have a way to provide that question to the user.