FS Write Scoping
Deciding on jumping all in on the file system! I wanted to be able to support both a JavaScript API for this, and a normal CLI setup, but trying to keep a JavaScript API setup (functionally-based) isn't as feasible with tools like FFmpeg, which primarily work and expect to work as a component on top of an available file system. When I tried to use FFmpeg with stdout, things got a bit hairy, at least in terms of Node.js Buffer objects they can only hold so much data, not as much as the file system can.