Tearing the CLI from the API
I want to reintroduce a Node-only API for using Art-Gen with JS too!

To do that though, I need to separate what is part of the CLI implementation from the tool itself. NBTify manages that nicely, but I didn't know how to do that when I initially set up this project.

I also want to make it a proper Node CLI, rather than one that needs to run with `tsx`. That also involves properly introducing the TS build step into the management of the project too.