CLI Args Demo
Looking into using a library for this again, it's still a little above my head though, in terms of understanding how to set it up, as well as get easy types out of it as well, which it doesn't completely seem to have here. I realized I think I want something like Zod, but for CLI args. Turns out someone made a project for that, so I'll probably use it next!

Here's all of the various concepts I'm coming back to this project with, some are existing concepts, some are the new CLI-related ones.

https://www.npmjs.com/package/@googleapis/youtube
https://googleapis.dev/nodejs/googleapis/latest/youtube/classes/Youtube.html
https://github.com/googleapis/google-api-nodejs-client

https://dev.to/rushankhan1/build-a-cli-with-node-js-4jbi
https://www.npmjs.com/package/yargs
https://www.reddit.com/r/node/comments/a1tl3i/what_library_should_i_use_for_creating_a_cli/
https://npmtrends.com/commander-vs-minimist-vs-nopt-vs-optimist-vs-optionator-vs-yargs
https://www.npmjs.com/package/commander
https://github.com/GregBrimble/zodcli

I wanted to complete this more, before committing it, but it's getting late, and I wanted to keep track of the links I found while reading up on this, and I think I don't quite understand how to use those libraries just yet, in order to get things working nice and properly.

Looking forward to using them for things like NBTify eventually too, it could really use some help with making things look and work nice! :D