CLI Removal (Temporary?)
I want to make a proper inclusive app build which will allow more non-coder kind of people to make use of the project, I think that's the current huge roadblock to making it something that can be more accessibly used.

I can add ffmpeg-static and Puppetter to an Electron app, and make a GUI with that itself, which will replace the CLI setup. Personally I really like the simplicity of the CLI, I think I need to make it more easy to use and install to actually make use of it though, since the CLI would require knowledge of using the terminal, installing Node.js, installing ffmpeg, already having Chrome installed, all just to actually use the project. Having a built GUI app is much easier to just plainly download, install, and run, and can have easier access for more people to check out. It's also much easier to showcase what it actually does as well, since it is just something you would be able to just download and install.

I will probably use Vite for developing the frontend, and either Solid or React for the UI itself. I'm curious about using SSR JSX for the image rendering template as well, rather than a custom-stringified HTML page, which is what it simply does now. Bundling the project all into one also allows the localhost server to just be it's own headless Electron window as well, which removes the need to start up a Node HTTP server just to make the thumbnail renderer accessible.

The concept for hooking all of that together feels fairly trivial, but it's a big hunk of tech to tie together and I want to make sure I do it right.

It does unfortunately make the project as a whole more complex because of the other dependencies it has to bundle together, but that in turn makes it much more accessible to the non-geeks like me that know how to set up all of those myself.