Recreational Outrage, Fix
Had to adjust the command pattern to work with one of the song names, it was because the RegExp I was using was being too catchy. This was from a week or two ago, didn't commit as it was just a temp fix, came back now to add it. So this was around August 11th or so.