Type Issues Fixes
This is because of the changes to TS itself, as well as the standard Node typings. It's a bummer how no longer a Buffer can be a Uint8Array. It was logical for a long time that a Uint8Array can't be a Buffer, because it was missing features only available in Node. But now both ways? Damn. Not a nice DX thing. I wish Node started using Uint8Array instead of Buffer for it's std functions. Too late to change that now of course though. And whatever they do, PLEASE don't modify Uint8Array to have Buffer methods. That would be even worse.