Big Formatting
Just general formatting from over the past few years, which I started to get over being super particular about wanting to format things "my way", or "the way I like them". It's just what you're used to. I see that now. Code.org didn't use spaces between commas and block scopes, so I learned I wasn't going to either. Who'da thunk? I format it with tools now because it is the "standardized" way to have the code presented, and easier to manage the code. It's not about "what I like" anymore. Or rather, the convenience of managing and diffing the code is more inline with "what I like" than the actual presentation.

And big picture, I actually think I prefer the look of it now too, after having gotten used to it. All of it is arbitrary though, it's just what you currently decide things to be. :)